Issue #10934: OutOfMemory Error following ANTLR4 upgrade
