+++ .ci/jsoref-spellchecker/whitelist.words
@@ -621,6 +621,7 @@ Izmailov
 jacoco
 Jakub
 jarchitect
+javaastvisitor
 javac
 javadoc
 javadocblocktaglocation
