+++ pom.xml
@@ -3482,7 +3482,6 @@
               </targetClasses>
               <targetTests>
                 <param>com.puppycrawl.tools.checkstyle.grammar.*</param>
-                <param>com.puppycrawl.tools.checkstyle.JavaParserTest</param>
                 <param>com.puppycrawl.tools.checkstyle.JavaAstVisitorTest</param>
               </targetTests>
               <excludedTestClasses>
